fix error

+++ src/main/java/com/quizptit/review/dto/ReviewSubjectDTO.java
@@ -11,12 +11,12 @@ public class ReviewSubjectDTO {
     private Long totalReviewCount;    // Số câu cần ôn
     private BigDecimal averageScore; // Điểm thông thạo trung bình
 
-    public ReviewSubjectDTO(Long subjectId, String subjectName, Long totalReviewCount, Double averageScore) {
+    public ReviewSubjectDTO(Long subjectId, String subjectName, Long totalReviewCount, BigDecimal averageScore) {
         this.subjectId = subjectId;
         this.subjectName = subjectName;
         this.totalReviewCount = totalReviewCount != null ? totalReviewCount : 0L;
         this.averageScore = (averageScore != null) 
-            ? BigDecimal.valueOf(averageScore).setScale(2, java.math.RoundingMode.HALF_UP) 
+            ? averageScore.setScale(2, java.math.RoundingMode.HALF_UP) 
             : BigDecimal.ZERO;
     }
 }
